[PLAT-3446] Fix null check in UniverseUpdateRootCert for rootCert privateKey
Summary:
In the diff D15980 a change to UniverseUpdateRootCert has been introduced to support k8s cert rotation.
Missed to add a null check for root cert private key as for custom certs private key will be null.
This diff adds the null which resolves the failing itests

Test Plan: Tested cert rotation manually for custom certs

Reviewers: vpatibandla

Reviewed By: vpatibandla

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D16128

+++ managed/src/main/java/com/yugabyte/yw/commissioner/tasks/subtasks/UniverseUpdateRootCert.java
@@ -139,7 +139,7 @@ public class UniverseUpdateRootCert extends UniverseTaskBase {
                 });
 
             rootCertFile.delete();
-            if (rootCert.privateKey.contains("ca.multi.root")) {
+            if (rootCert.privateKey != null && rootCert.privateKey.contains("ca.multi.root")) {
               File rootCertKey = new File(rootCert.privateKey);
               rootCertKey.delete();
             }
